cleanup
Also introducing a new revision of the clean up profile.

+++ src/main/java/com/shinoow/abyssalcraft/api/recipe/EngraverRecipes.java
@@ -110,7 +110,7 @@ public class EngraverRecipes {
 
 		return experienceList.entrySet().stream()
 				.filter(e -> areStacksEqual(stack, e.getKey()))
-				.map(e -> e.getValue().floatValue())
+				.map(e -> e.getValue())
 				.findFirst()
 				.orElse(0.0F);
 	}
